Digital.Abundance.Theory.& “DATAllowance”
My dream is to find a team. So we can build…
“DAT” (Digital. Abundance. theory) the system!
“DAT” will create “DATA” (Digital.Abundance.Theory.Allowance) for all people of the earth. 
If I was able to create what I can see.  Then you would know it’s changed the world, but that would be centralized thinking.  So I need to find many capable minds and investment so it can be decentralized and able to be built, implemented and watch how it grows.   This is the kind of project that will change how the world thinks and acts. It will stop all the lies of scarcity which enslaves all of our lives.  It will allow all to be free and to live in love and toil in things they love because the theory of abundance will be unleashed digitally, so all will soon be free from centralized lies. Truth will shine across the decentralized land, where every life matters!  Please embark on this giant mission. Whether you help build it. Or soon one day you will have a coin of “DATA” in your hand and you’ll know that it came from the “Digital.Abundance.Theory” team.